+++ src/application/NodeManager.ts
@@ -2,7 +2,11 @@ import type { PteroApp } from '.';
 import { BaseManager } from '../structures/BaseManager';
 import { Dict } from '../structures/Dict';
 import { Node }from '../structures/Node';
-import { CreateNodeOptions, NodeConfiguration } from '../common/app';
+import {
+    CreateNodeOptions,
+    NodeConfiguration,
+    NodeDeploymentOptions
+} from '../common/app';
 import {
     Filter,
     FilterArray,
@@ -60,8 +64,8 @@ export class NodeManager extends BaseManager {
         if (typeof obj === 'string') return this.cache.find(
             n => n.name === obj
         );
-        if (obj.relationships?.user)
-            return this._patch(obj.relationships.user) as Node;
+        if (obj.relationships?.node)
+            return this._patch(obj.relationships.node) as Node;
 
         return undefined;
     }
@@ -82,6 +86,14 @@ export class NodeManager extends BaseManager {
         const data = await this.client.requests.get(
             (id ? endpoints.nodes.get(id) : endpoints.nodes.main),
             options, this
+        );
+        return this._patch(data) as any;
+    }
+
+    /** compatibility issue with node-fetch */
+    private async fetchDeployable(options: NodeDeploymentOptions): Promise<Dict<number, Node>> {
+        const data = await this.client.requests._make(
+            endpoints.nodes.deploy, options, 'GET'
         );
         return this._patch(data) as any;
     }